refactor: ContextType와 StorageType을 기본 제공 어노테이션으로 전환
