+++ app/src/main/java/woowacourse/shopping/di/DiApplicationModule.kt
@@ -3,11 +3,13 @@ package woowacourse.shopping.di
 import android.content.Context
 import com.re4rk.arkdi.DiContainer
 import com.re4rk.arkdi.Singleton
+import com.re4rk.arkdi.annotations.ContextType
+import com.re4rk.arkdi.annotations.ContextType.Type.APPLICATION
+import com.re4rk.arkdi.annotations.StorageType
+import com.re4rk.arkdi.annotations.StorageType.Type.IN_MEMORY
 import woowacourse.shopping.data.CartInMemoryRepository
 import woowacourse.shopping.data.ProductSampleRepository
 import woowacourse.shopping.data.ShoppingDatabase
-import woowacourse.shopping.di.ContextType.Type.APPLICATION
-import woowacourse.shopping.di.StorageType.Type.IN_MEMORY
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
